refactor(optimum): drop the embed_tokens dtype casts
optimum-rbln loads embed_tokens in the compile dtype as of
RBLN-SW/optimum-rbln#674, so the lookup output no longer needs a cast to
rbln_config.dtype. Requires an optimum-rbln release that includes that PR.

+++ vllm_rbln/model_executor/models/optimum/qwen2_vl.py
@@ -398,9 +398,7 @@ class RBLNOptimumQwenVLForConditionalGeneration(
         """Embed the text tokens, then scatter each modality's features over
         its placeholder positions.
         """
-        inputs_embeds = self.model.embed_tokens(input_ids).to(
-            self.model.rbln_config.dtype
-        )
+        inputs_embeds = self.model.embed_tokens(input_ids)
         if not isinstance(multimodal_embeddings, dict) or not multimodal_embeddings:
             return inputs_embeds
         config = self.model.config
@@ -568,7 +566,7 @@ class RBLNOptimumQwenVLForConditionalGeneration(
             padded_batch_size = kwargs.pop("padded_batch_size", self.decoder_batch_size)
             self.model.decoder = self.model.decoders[padded_batch_size]
             input_ids = kwargs.pop("input_ids")
-            inputs_embeds = self.model.embed_tokens(input_ids).to(self.dtype)
+            inputs_embeds = self.model.embed_tokens(input_ids)
             logits = self.model.decoder(
                 inputs_embeds=inputs_embeds,
                 cache_position=cache_position,
